Rename "wiki" to "doc"
The name "wiki" is a holdover from many years ago when we used TracWiki.
It is not accurate to call the documentation a "wiki" and has not been
accurate for many years. In fact, it was never a public wiki that anyone
could edit. Everywhere else, we refer to the documentation with "doc,"
so it's time this directory caught up.

In a separate commit, I'll change all appropriate instances of "wiki" to
"doc" so that images continue to work correctly as before.